+++ src/documentation/release_notes.md
@@ -949,6 +949,11 @@ managers (backported to Jaybird 3.0.5) ([JDBC-468](http://tracker.firebirdsql.or
 -   Changed: The value returned by `ResultSetMetaData.getColumnDisplaySize` was 
 revised for `REAL`/`FLOAT` and `DOUBLE PRECISION` to take scientific notation 
 into account ([JDBC-514](http://tracker.firebirdsql.org/browse/JDBC-514))
+-   Fixed: Database metadata pattern parameters now allow the pattern escape 
+character (`\`) to occur unescaped, this means that patterns `A\B` and `A\\B` 
+will both match a value of `A\B`. This complies with the (undocumented) JDBC 
+expectation that patterns follow the ODBC requirements for pattern value 
+arguments ([JDBC-562](http://tracker.firebirdsql.org/browse/JDBC-562))
 
 Removal of deprecated classes and packages
 ------------------------------------------
@@ -1137,6 +1142,8 @@ The following methods will be removed in Jaybird 5:
     `MaintenanceManager.getLimboTransactions()` instead.
 -   `TraceManager.loadConfigurationFromFile(String)`, use standard Java 
     functionality like `new String(Files.readAllBytes(Paths.get(fileName)), <charset>)`
+-   `FBDatabaseMetaData.hasNoWildcards(String pattern)`
+-   `FBDatabaseMetaData.stripEscape(String pattern)`
     
 ### Removal of deprecated constants ###